Cooker CRUD

+++ src/main/java/com/cooker/cook/repositories/CookerRepository.java
@@ -3,5 +3,5 @@ package com.cooker.cook.repositories;
 import com.cooker.cook.entities.Cooker;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CookerRepository extends JpaRepository <Cooker,Integer> {
+public interface CookerRepository extends JpaRepository <Cooker,Long> {
 }
