+++ src/main/java/com/cooker/cook/repositories/IngridientRepository.java
@@ -3,5 +3,5 @@ package com.cooker.cook.repositories;
 import com.cooker.cook.entities.Ingridient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngridientRepository extends JpaRepository<Ingridient,Integer> {
+public interface IngridientRepository extends JpaRepository<Ingridient,Long> {
 }
